Add site and beginnings of api to tf
This code adds an s3 bucket for the js/html/css to be run from. This
will eventually be fronted with cloudfront. For now, I don't have enough
understanding of cloudfront to actually wire it all up.

This also adds the TF to upload the compiled lambda, and invoke it from
the aws api gateway.