shore up api gateway/lambda tf
In the api gateway, I read up a little more about how to configure it.
I don't want any path to be able to call this lambda function, I only
want /webhook to be able to call the function. This means I had to
refactor the lambda permission, the api resource, and the api method a
little.

I removed variables.tf because it wasn't being used.