Refactor and add some dev details to the readme
This is a refactoring that changes some of the tf naming to align with
my understanding of how things work. I extracted the things into
variables that will need to be changed should someone else ever want to
deploy this code to their own aws account.

I also came up with unique s3 bucket names for the site and the remote
state.