Hook up the image sync job queue
This adds an sqs queue for communication between the webhook and the
image sync function. It gives IAM permission to the sync function
to receive messages from the queue, and delete them from the queue.
I've added an event source resource mapping from the queue to the
image sync function, I believe this has the effect of launching
the function when messages are inbound into the queue.