Fix taint logic and add in dependencies
The deps from the virtualenv were not being added to the lambda zip
packages because I was trying to debug an error earlier on in the
project. Now that I've got most of the infrastructure in tf sorted
I wanted to get the boto library back into scope for my lambdas.

Also, when I refactored to use tf modules, it broke the taint logic
this commit fixes that logic so that taint is now working again in the
makefile.